+++ guillotina_s3storage/storage.py
@@ -9,15 +9,14 @@ from guillotina import configure
 from guillotina.browser import Response
 from guillotina.component import getUtility
 from guillotina.event import notify
+from guillotina.files import BaseCloudFile
+from guillotina.files import read_request_data
 from guillotina.interfaces import IAbsoluteURL
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IFileManager
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
-from guillotina.interfaces import IValueToJson
 from guillotina.schema import Object
-from guillotina.schema.fieldproperty import FieldProperty
-from guillotina.utils import get_content_path
 from guillotina.utils import get_current_request
 from guillotina_s3storage.events import FinishS3Upload
 from guillotina_s3storage.events import InitialS3Upload
@@ -34,7 +33,6 @@ import base64
 import boto3
 import botocore
 import logging
-import mimetypes
 import uuid
 
 
@@ -44,72 +42,11 @@ MAX_SIZE = 1073741824
 
 MIN_UPLOAD_SIZE = 5 * 1024 * 1024
 CHUNK_SIZE = MIN_UPLOAD_SIZE
-MAX_REQUEST_CACHE_SIZE = 6 * 1024 * 1024
 MAX_RETRIES = 5
-
-
-class UnRetryableRequestError(Exception):
-    pass
-
-
-async def read_request_data(request, chunk_size=CHUNK_SIZE):
-    if getattr(request, '_retry_attempt', 0) > 0:
-        # we are on a retry request, see if we have read cached data yet...
-        if request._retry_attempt > getattr(request, '_last_cache_data_retry_count', 0):
-            if request._cache_data is None:
-                # request payload was too large to fit into request cache.
-                # so retrying this request is not supported and we need to throw
-                # another error
-                raise UnRetryableRequestError()
-            data = request._cache_data[request._last_read_pos:request._last_read_pos + chunk_size]
-            request._last_read_pos += len(data)
-            if request._last_read_pos >= len(request._cache_data):
-                # done reading cache data
-                request._last_cache_data_retry_count = request._retry_attempt
-            return data
-
-    if not hasattr(request, '_cache_data'):
-        request._cache_data = b''
-
-    try:
-        data = await request.content.readexactly(chunk_size)
-    except asyncio.IncompleteReadError as e:
-        data = e.partial
-
-    if request._cache_data is not None:
-        if len(request._cache_data) + len(data) > MAX_REQUEST_CACHE_SIZE:
-            # we only allow caching up to chunk size, otherwise, no cache data..
-            request._cache_data = None
-        else:
-            request._cache_data += data
-
-    request._last_read_pos += len(data)
-    return data
 
 
 class S3Exception(Exception):
     pass
-
-
-def _to_str(value):
-    if isinstance(value, bytes):
-        value = value.decode('utf-8')
-    return value
-
-
-@configure.adapter(
-    for_=IS3File,
-    provides=IValueToJson)
-def json_converter(value):
-    if value is None:
-        return value
-    return {
-        'filename': value.filename,
-        'content_type': _to_str(value.content_type),
-        'size': value.size,
-        'extension': value.extension,
-        'md5': value.md5
-    }
 
 
 @configure.adapter(
@@ -129,10 +66,10 @@ class S3FileManager(object):
         """
         self.context._p_register()  # writing to object
 
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             file = S3File(content_type=self.request.content_type)
-            self.field.set(self.context, file)
+            self.field.set(self.field.context or self.context, file)
             # XXX no savepoint support right now?
             # trns = get_transaction(self.request)
             # trns.savepoint()
@@ -159,21 +96,21 @@ class S3FileManager(object):
         else:
             file.filename = uuid.uuid4().hex
 
-        await file.initUpload(self.context)
+        await file.init_upload(self.context)
         self.request._last_read_pos = 0
-        data = await read_request_data(self.request)
+        data = await read_request_data(self.request, CHUNK_SIZE)
 
         count = 0
 
         # If we have data or is an empty file
         while data or (len(data) == 0 and count == 0):
             old_current_upload = file._current_upload  # noqa
-            await file.appendData(data)
+            await file.append_data(data)
             count += 1
-            data = await read_request_data(self.request)
+            data = await read_request_data(self.request, CHUNK_SIZE)
 
         # Test resp and checksum to finish upload
-        await file.finishUpload(self.context)
+        await file.finish_upload(self.context)
 
     async def tus_create(self):
         self.context._p_register()  # writing to object
@@ -182,10 +119,10 @@ class S3FileManager(object):
         if self.request.headers.get('X-HTTP-Method-Override') == 'PATCH':
             return await self.tus_patch()
 
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             file = S3File(content_type=self.request.content_type)
-            self.field.set(self.context, file)
+            self.field.set(self.field.context or self.context, file)
         if 'CONTENT-LENGTH' in self.request.headers:
             file._current_upload = int(self.request.headers['CONTENT-LENGTH'])
         else:
@@ -218,7 +155,7 @@ class S3FileManager(object):
         else:
             file._one_tus_shoot = False
 
-        await file.initUpload(self.context)
+        await file.init_upload(self.context)
 
         # Location will need to be adapted on aiohttp 1.1.x
         resp = Response(headers={
@@ -232,7 +169,7 @@ class S3FileManager(object):
     async def tus_patch(self):
         self.context._p_register()  # writing to object
 
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if 'CONTENT-LENGTH' in self.request.headers:
             to_upload = int(self.request.headers['CONTENT-LENGTH'])
         else:
@@ -250,21 +187,21 @@ class S3FileManager(object):
             # One time shoot
             if file._block > 1:
                 raise AttributeError('You should push 5Mb blocks AWS')
-            await file.oneShotUpload(self.context, data)
+            await file.one_shot_upload(self.context, data)
             expiration = datetime.now() + timedelta(days=365)
         else:
             count = 0
             while data:
-                resp = await file.appendData(data)
+                resp = await file.append_data(data)
                 count += 1
 
-                data = await read_request_data(self.request)
+                data = await read_request_data(self.request, CHUNK_SIZE)
 
             expiration = file._resumable_uri_date + timedelta(days=7)
         if file._size <= file._current_upload:
-            await file.finishUpload(self.context)
+            await file.finish_upload(self.context)
         resp = Response(headers={
-            'Upload-Offset': str(file.actualSize()),
+            'Upload-Offset': str(file.get_actual_size()),
             'Tus-Resumable': '1.0.0',
             'Upload-Expires': expiration.isoformat(),
             'Access-Control-Expose-Headers': 'Upload-Offset,Upload-Expires,Tus-Resumable'
@@ -272,11 +209,11 @@ class S3FileManager(object):
         return resp
 
     async def tus_head(self):
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             raise KeyError('No file on this context')
         head_response = {
-            'Upload-Offset': str(file.actualSize()),
+            'Upload-Offset': str(file.get_actual_size()),
             'Tus-Resumable': '1.0.0',
             'Access-Control-Expose-Headers': 'Upload-Offset,Upload-Length,Tus-Resumable'
         }
@@ -297,7 +234,7 @@ class S3FileManager(object):
     async def download(self, disposition=None):
         if disposition is None:
             disposition = self.request.GET.get('disposition', 'attachment')
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             raise AttributeError('No field value')
 
@@ -324,7 +261,7 @@ class S3FileManager(object):
         return resp
 
     async def iter_data(self):
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             raise AttributeError('No field value')
 
@@ -342,61 +279,28 @@ class S3FileManager(object):
                         filename=None):
         self.context._p_register()  # writing to object
 
-        file = self.field.get(self.context)
+        file = self.field.get(self.field.context or self.context)
         if file is None:
             file = S3File(content_type=content_type)
-            self.field.set(self.context, file)
+            self.field.set(self.field.context or self.context, file)
 
         file._size = size
         if filename is None:
             filename = uuid.uuid4().hex
         file.filename = filename
 
-        await file.initUpload(self.context)
+        await file.init_upload(self.context)
 
         async for data in generator():
-            await file.appendData(data)
+            await file.append_data(data)
 
-        await file.finishUpload(self.context)
+        await file.finish_upload(self.context)
+        return file
 
 
 @implementer(IS3File)
-class S3File:
+class S3File(BaseCloudFile):
     """File stored in a GCloud, with a filename."""
-
-    filename = FieldProperty(IS3File['filename'])
-
-    def __init__(self, content_type='application/octet-stream',
-                 filename=None, size=0, md5=None):
-        if not isinstance(content_type, bytes):
-            content_type = content_type.encode('utf8')
-        self.content_type = content_type
-        if filename is not None:
-            self.filename = filename
-            extension_discovery = filename.split('.')
-            if len(extension_discovery) > 1:
-                self._extension = extension_discovery[-1]
-        elif self.filename is None:
-            self.filename = uuid.uuid4().hex
-
-        self._size = size
-        self._md5 = md5
-
-    def guess_content_type(self):
-        ct = _to_str(self.content_type)
-        if ct == 'application/octet-stream':
-            # try guessing content_type
-            ct, _ = mimetypes.guess_type(self.filename)
-            if ct is None:
-                ct = 'application/octet-stream'
-        return ct
-
-    def generate_key(self, request, context):
-        return '{}{}/{}::{}'.format(
-            request._container_id,
-            get_content_path(context),
-            context._p_oid,
-            uuid.uuid4().hex)
 
     async def copy_cloud_file(self, new_uri):
         if self.uri is None:
@@ -416,7 +320,7 @@ class S3File:
         await util._s3aioclient.delete_object(
             Bucket=self._bucket_name, Key=old_uri)
 
-    async def initUpload(self, context):
+    async def init_upload(self, context):
         """Init an upload.
 
         self._uload_file_id : temporal url to image beeing uploaded
@@ -456,7 +360,7 @@ class S3File:
             Key=self._upload_file_id,
             UploadId=self._mpu['UploadId'])
 
-    async def appendData(self, data):
+    async def append_data(self, data):
         part = await self._upload_part(data)
         self._multipart['Parts'].append({
             'PartNumber': self._block,
@@ -476,10 +380,10 @@ class S3File:
             UploadId=self._mpu['UploadId'],
             Body=data)
 
-    def actualSize(self):
+    def get_actual_size(self):
         return self._current_upload
 
-    async def finishUpload(self, context):
+    async def finish_upload(self, context):
         util = getUtility(IS3BlobStore)
         # It would be great to do on AfterCommit
         if self.uri is not None:
@@ -505,7 +409,7 @@ class S3File:
             UploadId=self._mpu['UploadId'],
             MultipartUpload=self._multipart)
 
-    async def oneShotUpload(self, context, data):
+    async def one_shot_upload(self, context, data):
         util = getUtility(IS3BlobStore)
 
         if hasattr(self, '_upload_file_id') and self._upload_file_id is not None:  # noqa
@@ -540,7 +444,7 @@ class S3File:
             self._upload_file_id)
         return response
 
-    async def deleteUpload(self, uri=None):
+    async def delete_upload(self, uri=None):
         util = getUtility(IS3BlobStore)
         if uri is None:
             uri = self.uri
@@ -565,40 +469,6 @@ class S3File:
     async def _download(self, url):
         util = getUtility(IS3BlobStore)
         return await util._s3aioclient.get_object(Bucket=self._bucket_name, Key=url)
-
-    def _set_data(self, data):
-        raise NotImplemented('Only specific upload permitted')
-
-    def _get_data(self):
-        raise NotImplemented('Only specific download permitted')
-
-    data = property(_get_data, _set_data)
-
-    @property
-    def uri(self):
-        if hasattr(self, '_uri'):
-            return self._uri
-
-    @property
-    def size(self):
-        if hasattr(self, '_size'):
-            return self._size
-        else:
-            return None
-
-    @property
-    def md5(self):
-        if hasattr(self, '_md5'):
-            return self._md5
-        else:
-            return None
-
-    @property
-    def extension(self):
-        if hasattr(self, '_extension'):
-            return self._extension
-        else:
-            return None
 
     @property
     def one_tus_shoot(self):